requirements.txt
schedule
bybit
pybit
ccxt==4.1.84
numpy==1.22.4
pandas==2.0.1
pandas-ta==0.3.14b0
python-dotenv==1.0.0
asyncio==3.4.3